requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.102.0
pydantic==2.11.7
pydantic_core==2.33.2
pylatexenc==2.10
python-dotenv==1.1.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
